+++ myqueue/test/test_mq.py
@@ -31,7 +31,6 @@ def test_fail(mq):
     mq('submit shell:echo+hello -d time@sleep+a')
     mq('submit shell:echo+hello2 -d shell:echo+hello')
     mq.wait()
-    mq.config.mpiexec = 'echo'
     mq('info')
     mq('info 1 -v')
     mq('ls -S t')