+++ docs/releasenotes.rst
@@ -9,8 +9,14 @@ Release notes
 Next release
 ============
 
-* Email notifications: `mq modify ... -N dA`.  See :ref:`modify` and
+* Email notifications: ``mq modify ... -N dA``.  See :ref:`modify` and
   :ref:`notifications`.
+* Use ``mq info`` to get information about your MyQueue installation:
+
+  * version
+  * location of the source code
+  * location of ``.myqueueu/`` folder
+  * configuration
 
 
 Version 21.4.2